# engine/openapi.yaml
openapi: 3.0.3
info:
  title: OpenPipe Workflow API
  version: 0.6.0
  description: API for discovering, validating, diagnosing, scaffolding, and running OpenPipe workflows/pipes.
servers:
  - url: http://127.0.0.1:8787
paths:
  /health:
    get:
      summary: Service health
      responses:
        '200':
          description: OK

  /runs:
    get:
      summary: List recent run reports
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Recent runs

  /runs/summary:
    get:
      summary: Summarize recent run outcomes
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Run summary

  /runs/trends:
    get:
      summary: Run trends grouped by UTC day
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Run trends

  /workflows/scaffold:
    post:
      summary: Scaffold a new workflow pipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [packId]
              properties:
                packId:
                  type: string
                baseDir:
                  type: string
      responses:
        '200':
          description: Scaffold completed
        '409':
          description: Pipe already exists

  /workflows:
    get:
      summary: List workflows
      parameters:
        - in: query
          name: details
          required: false
          schema:
            type: boolean
          description: Include workflow metadata summaries when true.
      responses:
        '200':
          description: Workflow list

  /workflows/{packId}:
    get:
      summary: Describe one workflow
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow metadata
        '404':
          description: Workflow not found

  /workflows/{packId}/validate:
    get:
      summary: Validate workflow graph/contracts before execution
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow is valid
        '422':
          description: Workflow validation failed

  /workflows/{packId}/doctor:
    get:
      summary: Diagnose workflow quality from validation and recent runs
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Workflow diagnosis

  /workflows/{packId}/run:
    post:
      summary: Run a workflow
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '200':
          description: Run completed
        '400':
          description: Invalid request
        '422':
          description: Workflow validation failed

  /pipes:
    get:
      summary: Alias of /workflows
      parameters:
        - in: query
          name: details
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Pipe list

  /pipes/{pipeId}:
    get:
      summary: Alias of /workflows/{packId}
      parameters:
        - in: path
          name: pipeId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pipe metadata

  /pipes/{pipeId}/validate:
    get:
      summary: Alias of /workflows/{packId}/validate
      parameters:
        - in: path
          name: pipeId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pipe valid
        '422':
          description: Pipe invalid

  /pipes/{pipeId}/doctor:
    get:
      summary: Alias of /workflows/{packId}/doctor
      parameters:
        - in: path
          name: pipeId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Pipe diagnosis

  /pipes/{pipeId}/run:
    post:
      summary: Alias of /workflows/{packId}/run
      parameters:
        - in: path
          name: pipeId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '200':
          description: Run completed

  /openapi.yaml:
    get:
      summary: OpenAPI spec
      responses:
        '200':
          description: OpenAPI YAML

components:
  schemas:
    RunRequest:
      type: object
      properties:
        runDir:
          type: string
        resumeRunDir:
          type: string
        maxSteps:
          type: integer
        dryRun:
          type: boolean
        injectDeviation:
          type: string
          enum:
            - requirements_mismatch
            - architecture_mismatch
            - implementation_bug
            - verification_gap
        inputs:
          type: object
          additionalProperties: true